requirements.txt
# Generated from pyproject.toml dependencies
# Python >=3.11 required

openai-agents>=0.1.0
openai>=1.0.0
google-api-python-client>=2.147.0
google-auth>=2.35.0
requests>=2.31.0
twilio>=9.0.0
backoff>=2.2.1
